migrating github to shared endpoint
