+++ pages/studentPortal/web/React/github.tsx
@@ -1,21 +1,25 @@
 import { useRouter } from "next/router";
-import React from "react";
 import LessonComponent, {
   LessonComponentData,
 } from "../../../../components/studentPortal/lessons/LessonComponent";
 
 import { Button } from "../../../../components/ui/Button";
+import { getGithubLessonData } from "../../../api/studentPortal/lessons/web/react/github";
 
-type LessonProps = {
+export type LessonProps = {
   lessonComponents: LessonComponentData[];
 };
 
 const Github = ({ lessonComponents }: LessonProps) => {
+
   const router = useRouter();
 
   const handleContinue = () => {
     router.push("/studentPortal/web/React/tailwindcss-gridflex");
   };
+
+  if (!lessonComponents) return <div>Loading...</div>;
+
   return (
     <>
       <div className="grid grid-cols-1 gap-8 px-4 pt-4 m-8 sm:px-12">
@@ -33,63 +37,12 @@ const Github = ({ lessonComponents }: LessonProps) => {
 };
 
 export async function getServerSideProps() {
-  const lessonComponents: LessonComponentData[] = [
-    {
-      component: "title",
-      text: "Deploying a Project on Github & Vercel",
-    },
-    {
-      component: "resource-list",
-      resources: [
-        {
-          title: "Github",
-          description:
-            "The official Github site.  Click 'View' to Login and create an account.",
-          image: "/images/lessons/github-mark.png",
-          link: "https://github.com/login",
-        },
-        {
-          title: "Github Desktop",
-          description:
-            "Click 'View' for the download links for Github Desktop.",
-          image: "https://desktop.github.com/images/desktop-icon.svg",
-          link: "https://desktop.github.com/",
-        },
-        {
-          title: "Vercel",
-          description:
-            "The official Vercel site.  Click 'View' to Login and create an account.",
-          image: "https://assets.vercel.com/image/upload/front/zeit/og.png",
-          link: "https://vercel.com/login",
-        },
-      ],
-    },
-    {
-      component: "description",
-      text: "Watch the video below and follow Lucky's steps to deploy your test assignment!",
-    },
-    {
-      component: "loom-video",
-      videoId: "afe5fda5eb4a485282a115efd5efb212",
-    },
-    {
-      component: "description",
-      text: "Saw your own confetti? Congratulations!! You're up and running.",
-    },
-
-    {
-      component: "submission",
-      codeSandboxTitle: "Submit your link below",
-      link: "",
-      placeholder: "Vercel Link",
-      assignmentId: "github-vercel",
-    },
-    {
-      component: "description",
-      text: "Still having trouble? No worries! Feel free to slack message a fellow classmate or book a coaching session through the portal via the Coaches tab.",
-    },
-  ];
+  const lessonComponents = getGithubLessonData();
   return { props: { lessonComponents } };
 }
 
 export default Github;
+
+
+
+
